example-training/src/requirements.txt
unsloth==2025.8.5
datasets==3.6.0
trl==0.21.0
tqdm==4.67.1
torch==2.7.1
accelerate==1.10.0
bitsandbytes==0.47.0
transformers==4.55.1
peft==0.17.0
PyYAML==6.0.2
numpy==2.2.6
wandb==0.17.9
pandas==2.2.3